separate api-resolver
